Fixed System Registry Bugs and Added Clang-Tidy (#110)
* Refactoring code to prepare for job system migration

* Code formatting to clang-format standards

* Refactor and removed references to global update to work and the uses of thread manager

* Updated code to clang-format standards

* Updated code to clang-format standards

* Updated code to clang-format standards

* Updated code to clang-format standards

* Dont panic, this commit is just involving renaming the naming enforced by clang-tidy. These are just renaming changes with minimal changes built on top of the removal of global update

* Not a large change, just fixing up some clang-format violations within these code changes

* Fixed system registry code to not crash when changing version to flecs/4.0.4. Updating system registry and made some changes to how it works

* Updated code to not violate clang-format coding standards

* Cleaned up system registry's implementation and removed unused code in its implementation

* github workflow updating clang-tidy to version 19

+++ src/atlas/core/update_handlers/timer.cpp
@@ -20,7 +20,7 @@ namespace atlas {
     }
 
     float timer::elapsed() {
-        return duration_cast<microseconds>(current_time() - m_stop_watch)
+        return (float)duration_cast<microseconds>(current_time() - m_stop_watch)
           .count();
     }
